Remove vim-plug
Rationale: This repository is intended to be used for my personal
configuration files. However, vim-plug is a) not a config file,
and b) not written by me. To legally keep vim-plug in this repo,
I would have to bundle the MIT license with it. This is not my
preferred license, so I decided to create a script (Neovim/install-
vim-plug.sh) to work around this licensing issue.